add rest crud apis start zk interface implementation

+++ src/main/java/com/bogdan/service/manager/parts/incoming/invoice/repository/IncomingInvoiceRepository.java
@@ -3,6 +3,6 @@ package com.bogdan.service.manager.parts.incoming.invoice.repository;
 import com.bogdan.service.manager.parts.incoming.invoice.model.IncomingInvoice;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IncomingInvoiceRepository extends CrudRepository<IncomingInvoice, Integer> {
+public interface IncomingInvoiceRepository extends CrudRepository<IncomingInvoice, Long> {
     // no custom queries
 }
